reverted changes to files where files where changes are not allowed, these files (or parts of them) must be the exact same as in the original physionet repo. moved helpercode.py
